Using resolved PokemonName slot value instead of raw value
Updating LookupByNameIntentHandler to use the first resolved value
for the PokemonName slot instead of the raw value that does not
necessarily match one of our expected values for the slot.
This will fix issues where the value contains an unexpected accent
character or some other inconsistency the causes data retrieval to fail.